Refactor sql_query_form functions to static methods
