+++ libraries/classes/Sql.php
@@ -1371,7 +1371,7 @@ EOT;
             $message = Message::getMessageForAffectedRows($num_rows);
 
             // Ok, here is an explanation for the !$is_select.
-            // The form generated by sql_query_form.lib.php
+            // The form generated by PhpMyAdmin\SqlQueryForm
             // and db_sql.php has many submit buttons
             // on the same form, and some confusion arises from the
             // fact that $message_to_show is sent for every case.
